requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
bcrypt==4.0.1
mysqlclient==2.2.0
Flask-Migrate==4.0.5